requirements.txt
appdirs==1.4.3
appnope==0.1.0
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
backcall==0.1.0
black==19.3b0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
decorator==4.4.0
dj-database-url==0.5.0
Django==2.2.1
django-cors-headers==3.0.2
django-heroku==0.3.1
django-tagging==0.4.6
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
docopt==0.6.2
entrypoints==0.3
flake8==3.7.7
gunicorn==19.9.0
idna==2.8
ipython==7.5.0
ipython-genutils==0.2.0
isort==4.3.20
jedi==0.13.3
jwt==0.6.1
lazy-object-proxy==1.4.1
mccabe==0.6.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
pipenv==2018.11.26
prompt-toolkit==2.0.9
psycopg2==2.8.3
ptpython==2.0.4
ptyprocess==0.6.0
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.4.0
PyJWT==1.7.1
pylint==2.3.1
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
traitlets==4.3.2
typed-ast==1.3.5
urllib3==1.25.3
virtualenv==16.6.2
virtualenv-clone==0.5.3
wcwidth==0.1.7
whitenoise==4.1.3
wrapt==1.11.1
